chore: split commitlint custom rules into per-rule directories
Move each commitlint-config custom rule into src/rules/<rule-name>/
with a co-located index.test.ts, so a rule's implementation and tests
live in one place. src/index.ts now only imports the rule modules and
composes the plugin callbacks and severities from a single array, so
adding a rule no longer touches two spots in index.ts.

Type each rule callback with a minimal Parsed input instead of the
commitlint Commit type. Commit is CommitBase & Record<string, string |
null>, whose index signature makes notes-array literals unassignable
and forces casts in tests; the minimal type removes the casts while
staying compatible with the plugin Rule type.

Refs #2329
