+++ packages/commitlint-config/src/rules/breaking-change-requires-bang/index.test.ts
@@ -1,0 +1,76 @@
+import lint from "@commitlint/lint";
+import { describe, expect, it } from "vitest";
+
+import { name, rule } from "./index.js";
+
+// breaking change を宣言するなら header に `!` が必須。footer だけの `BREAKING CHANGE:` は禁止。
+// GitHub の squash commit では footer が畳まれて見えず、prefix と実態がズレるため。
+// type は問わない（`chore!` も可）。組み込みの breaking-change-exclamation-mark (XNOR) とは違い、
+// `!` 単独は許可・footer 単独だけを弾く一方向の含意。
+
+const runRule = (parsed: Parameters<typeof rule>[0]) => rule(parsed);
+
+describe("breaking-change-requires-bang (unit)", () => {
+  it("breaking marker なしの通常コミットは通過する", () => {
+    const [valid] = runRule({ header: "feat: add foo", notes: [] });
+    expect(valid).toBe(true);
+  });
+
+  it("header に `!` があり breaking note もある (feat!) は通過する", () => {
+    const [valid] = runRule({
+      header: "feat!: drop node 18",
+      notes: [{ title: "BREAKING CHANGE", text: "drop node 18" }],
+    });
+    expect(valid).toBe(true);
+  });
+
+  it("header に `!` なしで footer だけ breaking は失敗する", () => {
+    const [valid] = runRule({
+      header: "feat: add foo",
+      notes: [{ title: "BREAKING CHANGE", text: "removed old api" }],
+    });
+    expect(valid).toBe(false);
+  });
+
+  it("BREAKING-CHANGE (ハイフン) note も検出する", () => {
+    const [valid] = runRule({
+      header: "fix: patch",
+      notes: [{ title: "BREAKING-CHANGE", text: "changed signature" }],
+    });
+    expect(valid).toBe(false);
+  });
+
+  it("header 先頭に空白があっても `!` を検出して通過する", () => {
+    // commitlint は header を trim せず rule に渡す。consumer が header-trim を無効化しても
+    // 偶発的な先頭空白で bang を見落とさないことを保証する。
+    const [valid] = runRule({
+      header: "  feat!: x",
+      notes: [{ title: "BREAKING CHANGE", text: "x" }],
+    });
+    expect(valid).toBe(true);
+  });
+});
+
+describe("breaking-change-requires-bang (integration via @commitlint/lint)", () => {
+  const rules = { [name]: [2, "always"] } as const;
+  // 本番は config-conventional の conventionalcommits preset で動く。@commitlint/lint の
+  // 既定 parser は `BREAKING-CHANGE` (ハイフン) を note 化しないため、note 検出を本番と
+  // 揃えるよう noteKeywords を明示する。
+  const opts = {
+    plugins: { local: { rules: { [name]: rule } } },
+    parserOpts: { noteKeywords: ["BREAKING CHANGE", "BREAKING-CHANGE"] },
+  };
+
+  it("`!` なしで BREAKING CHANGE footer だけのコミットを弾く", async () => {
+    const message = ["feat: add foo", "", "BREAKING CHANGE: removed old api"].join("\n");
+    const result = await lint(message, rules, opts);
+    expect(result.valid).toBe(false);
+    expect(result.errors.some((e) => e.name === name)).toBe(true);
+  });
+
+  it("`BREAKING-CHANGE` (ハイフン) footer だけのコミットも弾く", async () => {
+    const message = ["feat: add foo", "", "BREAKING-CHANGE: removed old api"].join("\n");
+    const result = await lint(message, rules, opts);
+    expect(result.valid).toBe(false);
+  });
+});
